// apps/website/src/types/content/content.ts
import { ReadTimeResults } from "reading-time";
import { Frontmatter } from "./frontmatter";
import { Slug } from "./slug";

export type Content<TMeta = unknown> = {
    frontmatter: Frontmatter<TMeta>;
    slug: Slug;
    readingTime: ReadTimeResults;
    contentFileRelativePath: string;
    content: string;
};